streaming: extract pure helpers from backend.py
Move source/conditioning helpers, the encode_cond_pair and
  blend_for_strength closures, and startup warmup out of
  demos/realtime_motion_graph_web/backend.py into acestep/streaming/.
  Behavior is byte-identical; backend.py keeps 1-line wrappers so all
  existing call sites are unchanged.

  - acestep/streaming/source.py: _try_load_sidecar, _decode_audio_msg,
    _load_known_fixture_waveform, _normalize_time_signature,
    _resolve_bpm_key_source. Sidecar-precedence rule preserved verbatim.
  - acestep/streaming/encode.py: encode_cond_pair(session, ...) and
    blend_for_strength extracted from closures; session is now an
    explicit parameter.
  - acestep/streaming/warmup.py: WARMUP_STATE, _WarmupWS,
    _load_warmup_audio_frame, run_startup_warmup. handle_client is
    injected as a parameter so the package stays free of demo imports
    (Phase 3 replaces the injection with a StreamingSession call).

  server.py reads WARMUP_STATE from the new module and passes
  handle_client into run_startup_warmup.
